--- week02/mission/dark-mode/src/원래 App.tsx
@@ -1,14 +1,0 @@
-import './App.css'
-import { ThemeProvider } from './context/ThemeProvider'
-import ContextPage from './context/ContextPage'
-
-function App() {
-
-  return (
-    <ThemeProvider>
-      <ContextPage />
-    </ThemeProvider>
-  )
-}
-
-export default App
